# externalized-compute/src/main/resources/liftie-swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Liftie provides K8s provisioning across cloud providers.",
        "title": "Liftie API",
        "contact": {
            "name": "Compute Platform Team",
            "url": "https://cloudera.slack.com/messages/CHGPU2EJK"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9999",
    "paths": {
...
        "/liftie/api/v1/cluster/{cluster-id}": {
            "get": {
                "description": "Get status of a K8s cluster",
                "consumes": [
                    "application/json",
                    "application/yaml"
                ],
                "produces": [
                    "application/json",
                    "application/yaml"
                ],
                "summary": "Get status of a K8s cluster",
                "operationId": "get-cluster-status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster cluster-id",
                        "name": "cluster-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.ClusterView"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github.infra.cloudera.com_liftie_liftie_pkg_common.StatusMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github.infra.cloudera.com_liftie_liftie_pkg_common.StatusMessage"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a managed K8s cluster",
                "consumes": [
                    "application/json",
                    "application/yaml"
                ],
                "produces": [
                    "application/json",
                    "application/yaml"
                ],
                "summary": "Update a managed K8s cluster",
                "operationId": "update-managed-cluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster cluster-id",
                        "name": "cluster-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update cluster request",
                        "name": "cluster",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/common.UpdateClusterRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/common.UpdateClusterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github.infra.cloudera.com_liftie_liftie_pkg_common.StatusMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github.infra.cloudera.com_liftie_liftie_pkg_common.StatusMessage"
                        }
                    },
                    "406": {
                        "description": "Not Acceptable",
                        "schema": {
                            "$ref": "#/definitions/common.ValidationResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github.infra.cloudera.com_liftie_liftie_pkg_common.StatusMessage"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a K8s cluster",
                "consumes": [
                    "application/json",
                    "application/yaml"
                ],
                "produces": [
                    "application/json",
                    "application/yaml"
                ],
                "summary": "Delete a K8s cluster",
                "operationId": "delete-cluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster cluster-id",
                        "name": "cluster-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.DeleteClusterResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github.infra.cloudera.com_liftie_liftie_pkg_common.StatusMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github.infra.cloudera.com_liftie_liftie_pkg_common.StatusMessage"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update a Registered K8s cluster",
                "consumes": [
                    "application/json",
                    "application/yaml"
                ],
                "produces": [
                    "application/json",
                    "application/yaml"
                ],
                "summary": "Update a Registered K8s cluster",
                "operationId": "update-registered-cluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster cluster-id",
                        "name": "cluster-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update cluster request",
                        "name": "cluster",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/common.UpdateRegisteredClusterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.UpdateRegisteredClusterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github.infra.cloudera.com_liftie_liftie_pkg_common.StatusMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github.infra.cloudera.com_liftie_liftie_pkg_common.StatusMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github.infra.cloudera.com_liftie_liftie_pkg_common.StatusMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.AWSCloudLogs": {
            "type": "object",
            "properties": {
                "autoScalingGroupActivityLogs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.AutoScalingGroupActivity"
                    }
                },
                "cloudFormationStackEvents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.StackEvent"
                    }
                },
                "cloudWatchLogs": {
                    "$ref": "#/definitions/common.CloudWatchLogsOutput"
                }
            }
        },
...
        "common.AwsVpcCni": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                }
            }
        },
        "common.AzureActivityLog": {
            "type": "object",
            "properties": {
                "correlationId": {
                    "type": "string"
                },
                "eventTimestamp": {
                    "type": "string"
                },
                "logLevel": {
                    "type": "string"
                },
                "operationId": {
                    "type": "string"
                },
                "operationName": {
                    "type": "string"
                },
                "resourceGroupName": {
                    "type": "string"
                },
                "resourceId": {
                    "type": "string"
                },
                "resourceType": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "statusMessage": {
                    "type": "string"
                }
            }
        },
...
        "common.AzureUpdateClusterSpec": {
            "type": "object",
            "properties": {
                "updateAksManagedIdentity": {
                    "description": "Updates the ManagedIdentity used by the AKS cluster",
                    "type": "boolean"
                }
            }
        },
...
        "common.CloudProviderDetails": {
            "type": "object",
            "properties": {
                "awsLogs": {
                    "$ref": "#/definitions/common.AWSCloudLogs"
                },
                "azureLogs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.AzureActivityLog"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "common.ClusterMetadata": {
            "type": "object",
            "required": [
                "env",
                "name"
            ],
            "properties": {
                "cluster_owner": {
                    "$ref": "#/definitions/common.ClusterOwner"
                },
                "cluster_type": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "env": {
                    "description": "must be crn",
                    "type": "string"
                },
                "experience_type": {
                    "type": "string"
                },
                "experimental": {
                    "$ref": "#/definitions/common.Experimental"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "owner_email": {
                    "type": "string"
                },
                "tenant": {
                    "type": "string"
                },
                "workloads": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "common.ClusterOwner": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "crn": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first-name": {
                    "type": "string"
                },
                "last-name": {
                    "type": "string"
                },
                "user-id": {
                    "type": "string"
                }
            }
        },
...
        "common.ClusterView": {
            "type": "object",
            "properties": {
                "api_end_point": {
                    "type": "string"
                },
                "bootstrapScript": {
                    "type": "string"
                },
                "ca_cert": {
                    "type": "string"
                },
                "cloud": {
                    "type": "object"
                },
                "cloudPlatform": {
                    "type": "string"
                },
                "clusterStateVersion": {
                    "type": "integer"
                },
                "cluster_id": {
                    "type": "string"
                },
                "cluster_owner": {
                    "$ref": "#/definitions/common.ClusterOwner"
                },
                "cluster_status": {
                    "$ref": "#/definitions/github.infra.cloudera.com_liftie_liftie_pkg_common.StatusMessage"
                },
                "cluster_type": {
                    "type": "string"
                },
                "created_time": {
                    "type": "string"
                },
                "dashboardUrl": {
                    "type": "string"
                },
                "deleted_time": {
                    "type": "string"
                },
                "env": {
                    "description": "Env is Deprecated: Use EnvName and EnvCrn.",
                    "type": "string"
                },
                "envCrn": {
                    "type": "string"
                },
                "envName": {
                    "type": "string"
                },
                "failedEvents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github.infra.cloudera.com_liftie_liftie_pkg_common.Event"
                    }
                },
                "helm": {
                    "$ref": "#/definitions/common.Helm"
                },
                "imageCatalog": {
                    "$ref": "#/definitions/common.ImageCatalog"
                },
                "image_pull_secrets": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "kind": {
                    "type": "string"
                },
                "kubernetes": {
                    "$ref": "#/definitions/common.K8sStats"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "nodeCount": {
                    "type": "integer",
                    "format": "int64"
                },
                "nodeImageUpgrade": {
                    "$ref": "#/definitions/common.NodeImageUpgrade"
                },
                "operation": {
                    "$ref": "#/definitions/common.OperationInfo"
                },
                "proxiedServices": {
                    "$ref": "#/definitions/common.ProxiedServices"
                },
                "proxyEndpoint": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "storage": {
                    "$ref": "#/definitions/common.StorageSpec"
                },
                "tenant": {
                    "type": "string"
                },
                "updated_time": {
                    "type": "string"
                },
                "upgrade": {
                    "$ref": "#/definitions/common.K8sUpgrade"
                },
                "workloads": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "common.DeleteClusterResponse": {
            "type": "object",
            "properties": {
                "cluster_status": {
                    "$ref": "#/definitions/github.infra.cloudera.com_liftie_liftie_pkg_common.StatusMessage"
                }
            }
        },
...
        "common.Deployments": {
            "type": "object",
            "properties": {
                "autoscaler": {
                    "$ref": "#/definitions/common.Autoscaler"
                },
                "awsVpcCni": {
                    "$ref": "#/definitions/common.AwsVpcCni"
                },
                "horizontal_pod_autoscaler": {
                    "$ref": "#/definitions/common.HorizontalPodAutoScaler"
                },
                "istio": {
                    "$ref": "#/definitions/common.RequestDeploymentProfileDetails"
                },
                "logging": {
                    "$ref": "#/definitions/common.Logging"
                },
                "metering": {
                    "$ref": "#/definitions/common.Metering"
                },
                "monitoring": {
                    "$ref": "#/definitions/common.Monitoring"
                },
                "nvidia": {
                    "$ref": "#/definitions/common.Nvidia"
                },
                "repository": {
                    "description": "repository for images",
                    "type": "string"
                },
                "telemetry_kms_key_ARN": {
                    "type": "string"
                },
                "yunikorn": {
                    "$ref": "#/definitions/common.YuniKorn"
                }
            }
        },
        "common.DockerConfig": {
            "type": "object",
            "required": [
                "server"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "description": "TODO Make Name binding required as soon as possible",
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "server": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "common.Experimental": {
            "type": "object",
            "properties": {
                "aksManagedIdentity": {
                    "type": "string"
                },
                "deleteClusterOnError": {
                    "type": "boolean"
                },
                "liftieInfraNodeRole": {
                    "description": "example: node.role.kubernetes.io/liftie-infra or foo.bar.baz/liftie-infra",
                    "type": "string"
                }
            }
        },
...
        "common.Helm": {
            "type": "object",
            "properties": {
                "upgrade": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "common.ImageCatalog": {
            "type": "object",
            "properties": {
                "crn": {
                    "description": "The catalog CRN (either CRN or name is required)",
                    "type": "string"
                },
                "name": {
                    "description": "The catalog name (either CRN or name is required)",
                    "type": "string"
                }
            }
        },
...
        "common.K8sStats": {
            "type": "object",
            "properties": {
                "platformProvider": {
                    "type": "string"
                },
                "platformVersion": {
                    "type": "string"
                },
                "version": {
                    "description": "Kubernetes version (default 1.14)",
                    "type": "string"
                }
            }
        },
        "common.K8sUpgrade": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "$ref": "#/definitions/common.K8sUpgradeStatus"
                }
            }
        },
        "common.K8sUpgradeStatus": {
            "type": "object",
            "properties": {
                "state": {
                    "type": "string"
                }
            }
        },
...
        "common.NodeImageAvailableUpgrade": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "string"
                },
                "instanceGroup": {
                    "type": "string"
                },
                "upgrades": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "common.NodeImageUpgrade": {
            "type": "object",
            "properties": {
                "available": {
                    "description": "Available shows if node image upgrades are available",
                    "type": "boolean"
                },
                "upgrades": {
                    "description": "Upgrades lists the currently used images and the new upgraded images",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.NodeImageAvailableUpgrade"
                    }
                }
            }
        },
...
        "common.OperationInfo": {
            "type": "object",
            "properties": {
                "last": {
                    "type": "integer",
                    "format": "int64"
                },
                "last_successful": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
...
        "common.ProxiedService": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
        "common.ProxiedServices": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/common.ProxiedService"
            }
        },
...
        "common.RegisterClusterSpec": {
            "type": "object",
            "required": [
                "dockerConfigs",
                "k8sProvider",
                "kubeConfig"
            ],
            "properties": {
                "deployments": {
                    "$ref": "#/definitions/common.Deployments"
                },
                "dockerConfigs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.DockerConfig"
                    }
                },
                "k8sProvider": {
                    "$ref": "#/definitions/common.K8sProvider"
                },
                "kubeConfig": {
                    "$ref": "#/definitions/common.KubeConfig"
                }
            }
        },
...
        "common.Security": {
            "type": "object",
            "properties": {
                "apiServer": {
                    "$ref": "#/definitions/common.APIServer"
                },
                "private": {
                    "type": "boolean"
                },
                "secretEncryption": {
                    "$ref": "#/definitions/common.SecretEncryption"
                },
                "volumeEncryption": {
                    "$ref": "#/definitions/common.VolumeEncryption"
                }
            }
        },
...
        "common.StorageSpec": {
            "type": "object",
            "properties": {
                "csiEnabled": {
                    "type": "boolean"
                }
            }
        },
        "common.UpdateClusterDeployments": {
            "type": "object",
            "properties": {
                "autoscaler": {
                    "$ref": "#/definitions/common.Autoscaler"
                }
            }
        },
        "common.UpdateClusterMetadata": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "experimental": {
                    "$ref": "#/definitions/common.Experimental"
                }
            }
        },
        "common.UpdateClusterRequest": {
            "type": "object",
            "required": [
                "spec"
            ],
            "properties": {
                "clusterStateVersion": {
                    "type": "integer",
                    "format": "int64"
                },
                "metadata": {
                    "$ref": "#/definitions/common.UpdateClusterMetadata"
                },
                "skipValidation": {
                    "type": "boolean"
                },
                "spec": {
                    "$ref": "#/definitions/common.UpdateClusterSpec"
                }
            }
        },
        "common.UpdateClusterResponse": {
            "type": "object",
            "properties": {
                "clusterId": {
                    "type": "string"
                },
                "clusterStatus": {
                    "$ref": "#/definitions/github.infra.cloudera.com_liftie_liftie_pkg_common.StatusMessage"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
        "common.UpdateClusterSpec": {
            "type": "object",
            "properties": {
                "azure": {
                    "$ref": "#/definitions/common.AzureUpdateClusterSpec"
                },
                "deployments": {
                    "$ref": "#/definitions/common.UpdateClusterDeployments"
                },
                "helm": {
                    "$ref": "#/definitions/common.Helm"
                },
                "imageCatalog": {
                    "$ref": "#/definitions/common.ImageCatalog"
                },
                "refreshInstanceGroups": {
                    "type": "boolean"
                },
                "security": {
                    "$ref": "#/definitions/common.Security"
                },
                "storage": {
                    "$ref": "#/definitions/common.StorageSpec"
                }
            }
        },
...
        "common.UpdateRegisteredClusterRequest": {
            "type": "object",
            "required": [
                "apiVersion",
                "kind",
                "metadata",
                "spec"
            ],
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/common.ClusterMetadata"
                },
                "spec": {
                    "$ref": "#/definitions/common.RegisterClusterSpec"
                }
            }
        },
        "common.UpdateRegisteredClusterResponse": {
            "type": "object",
            "properties": {
                "cluster_id": {
                    "type": "string"
                },
                "cluster_status": {
                    "$ref": "#/definitions/github.infra.cloudera.com_liftie_liftie_pkg_common.StatusMessage"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
...
        "common.ValidationResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                },
                "summary": {
                    "$ref": "#/definitions/common.ValidationSummary"
                },
                "validations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.ValidationResult"
                    }
                }
            }
        },
        "common.ValidationResult": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "cloudProviders": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "detailedMessage": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "common.ValidationSummary": {
            "type": "object",
            "properties": {
                "failed": {
                    "type": "integer"
                },
                "passed": {
                    "type": "integer"
                },
                "skipped": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "warning": {
                    "type": "integer"
                }
            }
        },
...
        "github.infra.cloudera.com_liftie_liftie_pkg_common.Event": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "progress": {
                    "type": "string"
                },
                "resourceType": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "statusReason": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "github.infra.cloudera.com_liftie_liftie_pkg_common.StatusMessage": {
            "type": "object",
            "properties": {
                "cloudProviderDetails": {
                    "$ref": "#/definitions/common.CloudProviderDetails"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}